+++ apibuilder4elastic/custom_flow_nodes/api-builder-plugin-axway-api-management/test/test-customProperties-MappingMerge.js
@@ -5,13 +5,24 @@ const path = require('path');
 const fs = require('fs');
 const nock = require('nock');
 const envLoader = require('dotenv');
+const decache = require('decache');
 
 describe('Merge custom properties tests', () => {
 
 	let plugin;
 	let flowNode;
+
+	const envFilePath = path.join(__dirname, '.env');
+	if (fs.existsSync(envFilePath)) {
+		delete process.env.API_MANAGER; // Otherwise it is not overwritten
+		envLoader.config({ path: envFilePath });
+	}
+	// Delete the cached module 
+	decache('../config/axway-api-utils.default.js');
+	var pluginConfig = require('../config/axway-api-utils.default.js').pluginConfig['api-builder-plugin-axway-api-management'];
+
 	beforeEach(async () => {
-		plugin = await MockRuntime.loadPlugin(getPlugin, {MOCK_LOOKUP_API:"true"});
+		plugin = await MockRuntime.loadPlugin(getPlugin, {MOCK_LOOKUP_API:"true", ...pluginConfig});
 		plugin.setOptions({ validateOutputs: true });
 		flowNode = plugin.getFlowNode('axway-api-management');
 	});