+++ apibuilder4elastic/test/trafficMonitorAPI/authZConfig/authorization-config-skipped.js
@@ -1,4 +1,5 @@
 var authorizationConfig = {
+	enableUserAuthorization: false,
 	// For how long should the information cached by the API-Builder process
 	cacheTTL: parseInt(process.env.EXT_AUTHZ_CACHE_TTL) ? process.env.EXT_AUTHZ_CACHE_TTL : 300,
 	externalHTTP : {